review: docs and cleaning
