+++ tailwind.common.js
@@ -3,7 +3,7 @@
 module.exports = {
   theme: {
     fontFamily: {
-      lato: ['"Lato"', '-apple-system', 'sans-serif'],
+      sans: ['"Lato"', '-apple-system', 'sans-serif'],
     },
     colors: {
       // Make sure these guidelines are followed when adding new colors: https://tailwindcss.com/docs/customizing-colors#using-css-variables
@@ -48,7 +48,7 @@ module.exports = {
       // This diverges from tailwind's standard pattern of defining its own spacing numbers
       // but should be easier for developers trying to match spacing to designs in Figma.
       0: '0',
-      1: '0.0625rem', // 3px
+      1: '0.0625rem', // 1px
       3: '0.1875rem', // 3px
       5: '0.3125rem', // 5px
       10: '0.625rem', // 10px
@@ -57,6 +57,7 @@ module.exports = {
       25: '1.5625rem', // 25px
       30: '1.875rem', // 30px
       40: '2.5rem', // 40px
+      45: '2.8125rem', // 45px
       50: '3.125rem', // 50px
       60: '3.75rem', // 60px
     },
